ci: deploy executables to gh packages
