+++ apps/website/ci/start.sh
@@ -1,0 +1,6 @@
+#!/usr/bin/env sh
+
+set -o errexit
+set -o nounset
+
+npx remix-serve ./index.js
