+++ apps/website/package.json
@@ -15,10 +15,10 @@
     "@nextui-org/react": "1.0.0-beta.10",
     "@react-stately/collections": "3.4.4",
     "@react-stately/menu": "3.4.2",
-    "@remix-run/node": "1.7.4",
-    "@remix-run/react": "1.7.4",
-    "@remix-run/serve": "1.7.4",
-    "@remix-run/server-runtime": "1.7.4",
+    "@remix-run/node": "1.7.6",
+    "@remix-run/react": "1.7.6",
+    "@remix-run/serve": "1.7.6",
+    "@remix-run/server-runtime": "1.7.6",
     "@routes-gen/remix": "0.3.5",
     "@types/uuid": "8.3.4",
     "@uiw/react-codemirror": "4.12.3",
@@ -43,14 +43,11 @@
     "zodix": "0.1.0"
   },
   "devDependencies": {
-    "@remix-run/dev": "1.7.4",
-    "@remix-run/eslint-config": "1.7.4",
+    "@remix-run/dev": "1.7.6",
+    "@remix-run/eslint-config": "1.7.6",
     "@types/lodash": "4.14.185",
     "@types/luxon": "3.0.1",
     "@types/react": "18.0.20",
     "@types/react-dom": "18.0.6"
-  },
-  "scripts": {
-    "postinstall": "npx remix setup node"
   }
 }
