+++ pyproject.toml
@@ -6,7 +6,7 @@ authors = ["Diptesh Choudhuri <ignisda2001@gmail.com>"]
 readme = "README.md"
 
 [tool.poetry.dependencies]
-python = "^3.10"
+python = ">=3.8.1"
 jinja2 = "^3.1.2"
 click = "^8.1.3"
